Global variables and locks
tuple_space: This is a dictionary used to store key-value pairs (i.e., tuple space). The keys are unique strings, and the values are also strings.
lock: This is a thread lock used to synchronize access to tuple_space. Since the server is multi-threaded, multiple clients may attempt to read from or modify the tuple space simultaneously, so a lock is needed to prevent data races and inconsistencies.