Update server.py.txt
msg_len: The first part of the request represents the total length of the message.
command: The second part of the request indicates the command type (R, G or P).
key: The third part of the request represents the key.
value: If the request contains more parts, concatenate the remaining parts into values; Otherwise, the value is None.