Update server.py
My problem lies in that in Python, functions must be defined before being called. If the handle_client function is not defined before it is called, an error will occur. Furthermore, if handle_client is defined as a class method (such as a static method or an instance method), and I attempt to pass it as an ordinary function to threading.Thread, it will also cause problems.