Update Client.py
My problem lies in that in the current code, client_socket.close() is placed inside the for loop, which means that the client will close the connection immediately after sending each request instead of after all requests have been sent. This is inconsistent with the job requirements because the client should close the connection only after completing all the requests.