Update test_server_client.py
Set the port number that the server listens to.
Call the run_server function to start the server and use time.sleep(2) to pause for 2 seconds to ensure that the server has started and is ready to accept the connection.
Create a list of client request files, including files from client_1.txt to client_10.txt.
Traverse the list of client request files and call the run_client function for each file to start a client process.
Use client.wait() to wait for all client processes to complete.
Use server_process.terminate() to stop the server process, and use server_process.wait() to wait for the server process to stop completely.